+++ src/lib/services/ls.js
@@ -2,9 +2,11 @@ const path = require('path')
 const globSync = require('glob').globSync
 
 class Ls {
-  constructor (directory = null) {
-    this.cwd = directory ? path.resolve(directory) : './'
+  constructor (directory = './', envFile = '.env*') {
     this.ignore = ['node_modules/**', '.git/**']
+
+    this.cwd = path.resolve(directory)
+    this.envFile = envFile
   }
 
   run () {
@@ -17,7 +19,22 @@ class Ls {
       cwd: this.cwd // context dirctory for globSync
     }
 
-    return globSync('**/.env*', options)
+    const patterns = this._patterns()
+    return globSync(patterns, options)
+  }
+
+  _patterns () {
+    if (!Array.isArray(this.envFile)) {
+      return `**/${this.envFile}`
+    }
+
+    const out = []
+
+    for (let i = 0; i < this.envFile.length; i++) {
+      const part = this.envFile[i]
+      out.push(`**/${part}`)
+    }
+    return out
   }
 }